integracion de todas las pruebas en el archivo test_voraz
