--- .github/workflows/ignorar-pruebas.yml
@@ -1,33 +1,0 @@
-name: Ignorar pruebas
-
-on:
-  workflow_dispatch:
-  push:
-    branches: [ main ]
-    paths:
-      - ".github/workflows/ignorar-pruebas.yml"
-
-permissions:
-  contents: read
-
-concurrency:
-  group: ignorar-pruebas-${{ github.ref }}
-  cancel-in-progress: true
-
-jobs:
-  no-tests:
-    runs-on: ubuntu-latest
-    timeout-minutes: 5
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-
-      - name: Mensaje de política
-        run: |
-          echo "⚠️ Según el enunciado, las pruebas NO se ejecutan en CI por la licencia gratuita."
-          echo "✅ Las pruebas deben ejecutarse localmente y documentarse en el informe con resultados."
-          echo "ℹ️ Pipeline minimalista para evitar consumo excesivo de minutos."
-
-      - name: Verificación sintáctica rápida (sin tests)
-        run: |
-          python -m py_compile $(git ls-files 'src/*.py')
