+++ tests/benchmarks/bench_10.csv
@@ -24,3 +24,73 @@ n,repeticion,tiempo_segundos
 10,3,0.054605
 10,4,0.059286
 10,5,0.055682
+10,1,0.063532
+10,2,0.063370
+10,3,0.063118
+10,4,0.068705
+10,5,0.067774
+10,1,0.067673
+10,2,0.060878
+10,3,0.062620
+10,4,0.062732
+10,5,0.061931
+10,1,0.065348
+10,2,0.062552
+10,3,0.066078
+10,4,0.060563
+10,5,0.079849
+10,1,0.062074
+10,2,0.065595
+10,3,0.058654
+10,4,0.059148
+10,5,0.063233
+10,1,0.071184
+10,2,0.066877
+10,3,0.064561
+10,4,0.060817
+10,5,0.071740
+10,1,0.073595
+10,2,0.066972
+10,3,0.066341
+10,4,0.063699
+10,5,0.062633
+10,1,0.067145
+10,2,0.065873
+10,3,0.066132
+10,4,0.085495
+10,5,0.063475
+10,1,0.066810
+10,2,0.061365
+10,3,0.059310
+10,4,0.065720
+10,5,0.060045
+10,1,0.090125
+10,2,0.104612
+10,3,0.147293
+10,4,0.086809
+10,5,0.068393
+10,1,0.067802
+10,2,0.069456
+10,3,0.062958
+10,4,0.067838
+10,5,0.066578
+10,1,0.060615
+10,2,0.060117
+10,3,0.056959
+10,4,0.058372
+10,5,0.057832
+10,1,0.065011
+10,2,0.069096
+10,3,0.061245
+10,4,0.089755
+10,5,0.065257
+10,1,0.064269
+10,2,0.062168
+10,3,0.059322
+10,4,0.060648
+10,5,0.066506
+10,1,0.068939
+10,2,0.068158
+10,3,0.071375
+10,4,0.068541
+10,5,0.072168
