+++ tests/benchmarks/bench_100.csv
@@ -24,3 +24,73 @@ n,repeticion,tiempo_segundos
 100,3,0.060159
 100,4,0.061174
 100,5,0.062773
+100,1,0.093943
+100,2,0.083655
+100,3,0.071018
+100,4,0.059318
+100,5,0.068486
+100,1,0.065149
+100,2,0.063709
+100,3,0.064581
+100,4,0.061853
+100,5,0.060974
+100,1,0.063613
+100,2,0.081442
+100,3,0.060132
+100,4,0.061180
+100,5,0.061748
+100,1,0.062898
+100,2,0.059988
+100,3,0.064106
+100,4,0.066892
+100,5,0.058260
+100,1,0.066382
+100,2,0.061614
+100,3,0.065315
+100,4,0.065188
+100,5,0.077204
+100,1,0.091940
+100,2,0.074527
+100,3,0.062951
+100,4,0.061088
+100,5,0.060498
+100,1,0.068723
+100,2,0.063818
+100,3,0.078296
+100,4,0.087752
+100,5,0.086385
+100,1,0.063366
+100,2,0.063711
+100,3,0.057965
+100,4,0.062881
+100,5,0.063160
+100,1,0.073235
+100,2,0.069575
+100,3,0.060971
+100,4,0.061021
+100,5,0.064528
+100,1,0.065618
+100,2,0.066919
+100,3,0.065150
+100,4,0.086102
+100,5,0.066124
+100,1,0.061052
+100,2,0.059635
+100,3,0.061682
+100,4,0.058398
+100,5,0.061604
+100,1,0.066396
+100,2,0.068952
+100,3,0.059297
+100,4,0.070625
+100,5,0.067625
+100,1,0.065593
+100,2,0.061613
+100,3,0.060432
+100,4,0.062511
+100,5,0.085865
+100,1,0.081396
+100,2,0.065568
+100,3,0.060539
+100,4,0.064100
+100,5,0.078457
