+++ tests/benchmarks/bench_1000.csv
@@ -24,3 +24,73 @@ n,repeticion,tiempo_segundos
 1000,3,0.062080
 1000,4,0.063091
 1000,5,0.060549
+1000,1,0.070523
+1000,2,0.064223
+1000,3,0.064143
+1000,4,0.066380
+1000,5,0.062672
+1000,1,0.068447
+1000,2,0.065717
+1000,3,0.061737
+1000,4,0.063569
+1000,5,0.063941
+1000,1,0.068505
+1000,2,0.064140
+1000,3,0.062622
+1000,4,0.062737
+1000,5,0.064615
+1000,1,0.062758
+1000,2,0.060990
+1000,3,0.066910
+1000,4,0.060221
+1000,5,0.064145
+1000,1,0.073394
+1000,2,0.064096
+1000,3,0.064054
+1000,4,0.068363
+1000,5,0.066681
+1000,1,0.123893
+1000,2,0.092358
+1000,3,0.092320
+1000,4,0.073086
+1000,5,0.065843
+1000,1,0.065274
+1000,2,0.069503
+1000,3,0.065874
+1000,4,0.065809
+1000,5,0.070762
+1000,1,0.063629
+1000,2,0.099436
+1000,3,0.070118
+1000,4,0.066595
+1000,5,0.088778
+1000,1,0.073865
+1000,2,0.065030
+1000,3,0.061035
+1000,4,0.059799
+1000,5,0.060012
+1000,1,0.067511
+1000,2,0.063604
+1000,3,0.068723
+1000,4,0.070035
+1000,5,0.068022
+1000,1,0.072779
+1000,2,0.064089
+1000,3,0.066188
+1000,4,0.065476
+1000,5,0.062778
+1000,1,0.065547
+1000,2,0.066013
+1000,3,0.066074
+1000,4,0.063141
+1000,5,0.067005
+1000,1,0.066132
+1000,2,0.068069
+1000,3,0.065104
+1000,4,0.065473
+1000,5,0.059576
+1000,1,0.096765
+1000,2,0.068650
+1000,3,0.070992
+1000,4,0.067585
+1000,5,0.065125
