+++ tests/benchmarks/bench_10000.csv
@@ -24,3 +24,58 @@ n,repeticion,tiempo_segundos
 10000,3,0.082001
 10000,4,0.099251
 10000,5,0.079161
+10000,1,0.099180
+10000,2,0.100669
+10000,3,0.103392
+10000,4,0.099834
+10000,5,0.110986
+10000,1,0.099452
+10000,2,0.100128
+10000,3,0.103065
+10000,4,0.101255
+10000,5,0.096333
+10000,1,0.103218
+10000,2,0.123232
+10000,3,0.097488
+10000,4,0.156586
+10000,5,0.108468
+10000,1,0.111536
+10000,2,0.116684
+10000,3,0.118049
+10000,4,0.099810
+10000,5,0.102503
+10000,1,0.101683
+10000,2,0.099536
+10000,3,0.098450
+10000,4,0.100155
+10000,5,0.098208
+10000,1,0.098426
+10000,2,0.103241
+10000,3,0.096839
+10000,4,0.098988
+10000,5,0.093285
+10000,1,0.106243
+10000,2,0.101562
+10000,3,0.100639
+10000,4,0.121259
+10000,5,0.099480
+10000,1,0.101273
+10000,2,0.096988
+10000,3,0.095718
+10000,4,0.101350
+10000,5,0.094908
+10000,1,0.122389
+10000,2,0.099517
+10000,3,0.099867
+10000,4,0.101123
+10000,5,0.098969
+10000,1,0.097739
+10000,2,0.104986
+10000,3,0.090092
+10000,4,0.100019
+10000,5,0.096280
+10000,1,0.111492
+10000,2,0.112646
+10000,3,0.098580
+10000,4,0.101412
+10000,5,0.098103
