+++ tests/benchmarks/bench_50000.csv
@@ -24,3 +24,58 @@ n,repeticion,tiempo_segundos
 50000,3,0.211998
 50000,4,0.240827
 50000,5,0.208029
+50000,1,0.264960
+50000,2,0.265262
+50000,3,0.220252
+50000,4,0.220784
+50000,5,0.215501
+50000,1,0.232686
+50000,2,0.224724
+50000,3,0.230077
+50000,4,0.220151
+50000,5,0.224497
+50000,1,0.234646
+50000,2,0.247275
+50000,3,0.313833
+50000,4,0.276434
+50000,5,0.314554
+50000,1,0.276184
+50000,2,0.266974
+50000,3,0.238288
+50000,4,0.273598
+50000,5,0.237251
+50000,1,0.222161
+50000,2,0.259712
+50000,3,0.234649
+50000,4,0.250629
+50000,5,0.235235
+50000,1,0.219156
+50000,2,0.242485
+50000,3,0.220238
+50000,4,0.223747
+50000,5,0.240076
+50000,1,0.222013
+50000,2,0.221721
+50000,3,0.216680
+50000,4,0.221547
+50000,5,0.216278
+50000,1,0.214240
+50000,2,0.218147
+50000,3,0.302583
+50000,4,0.223953
+50000,5,0.210027
+50000,1,0.214676
+50000,2,0.214647
+50000,3,0.221689
+50000,4,0.213910
+50000,5,0.210725
+50000,1,0.222647
+50000,2,0.239243
+50000,3,0.227229
+50000,4,0.224252
+50000,5,0.258049
+50000,1,0.254418
+50000,2,0.220800
+50000,3,0.230403
+50000,4,0.231762
+50000,5,0.231587
